Refatoração do flow de autenticação.
